# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Marvol Facility Cleaning Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff members
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaffMember
      tags: [staff]
      summary: Create a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffMemberRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    put:
      operationId: updateStaffMember
      tags: [staff]
      summary: Update a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffMemberRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    delete:
      operationId: deleteStaffMember
      tags: [staff]
      summary: Delete a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    InternalCronSecret:
      type: apiKey
      in: header
      name: X-Internal-Cron-Secret
    SendGridInboundSecret:
      type: apiKey
      in: header
      name: X-SendGrid-Inbound-Secret
  schemas:
...
    StaffMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
          enum: [staff, supervisor, admin, inspector]
        hasEmail:
          type: boolean
          description: Whether the staff member has a configured sign-in email. The email address itself is never returned.
        active:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - role
        - hasEmail
        - active
        - createdAt
    CreateStaffMemberRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [staff, supervisor, admin, inspector]
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
      required:
        - name
        - role
    UpdateStaffMemberRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [staff, supervisor, admin, inspector]
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        active:
          type: boolean
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType, purpose]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
        purpose:
          type: string
          enum: [task_before, task_after, issue_before, issue_after, conversation_attachment, shared_photo, application_document]
        taskId:
          type: integer
        conversationId:
          type: integer
        issueId:
          type: integer
        areaId:
          type: integer
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        uploadToken:
          type: string
          format: uuid
          description: Applicant capability returned only for application_document uploads.
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...